Flatten pkgs
