+++ pkg/betterlogger/contexthandler.go
@@ -8,7 +8,6 @@ import (
 )
 
 // ContextHandler wraps a slog.Handler (e.g., JSON handler) and processes slogFields from the context.
-// Deprecated: Use the logging system provided in tracing instead.
 type ContextHandler struct {
 	innerHandler slog.Handler
 }
